Revert "Add draining for cc_uploader (#529)"
This reverts commit 45d2a21fc5e4d71e3dcf971b006e5a6d2b2b94db.

This original commit introduces the requirement that cc_uploader is
co-located on the same VM as the cloud_controller_ng job. While this is
the case in cf-deployment, this is not the case for all deployments of
CF (like the VMware one).

As this commit has broken our deployment, and its not easy for us simply
to relocate the job, we propose we revert this and discuss options going forwards.

The cc_uploader originally was designed an independent process, typically processes
that require to be co-located are within the same job.
E.g. nginx, cloud_controller_ng and cloud_controller_local_worker.

Some options:
* Investigate a way to generically drain uploads from any source in
  nginx rather than specifically in cc_uploader.
    * Current timeout is [10 seconds](https://github.com/cloudfoundry/cloud_controller_ng/blob/5c4dac049bd28979284aeab1efa48c7075676131/lib/cloud_controller/drain.rb#L9)
* Find a way to sync draining between jobs without requiring
  co-location.
* Keep the co-location requirement for draining but behind a capi-property.